TINY-7191: Made general improvements to PR
